Revert "增加一个 ddl"
This reverts commit ac33658d62cb3f38f3f83632add0e7a15d160abe.

+++ biz-shared/src/main/java/com/magicliang/transaction/sys/biz/shared/request/acceptance/AcceptanceCommand.java
@@ -31,7 +31,7 @@ public class AcceptanceCommand extends HandlerRequest {
      *
      * @see TransFundAccountingEntryTypeEnum
      */
-    private Integer accountingEntry;
+    private Short accountingEntry;
 
     /**
      * 支付备注，上游系统选填
